src/streamlit_ui/pyproject.toml
[project]
name = "learn-ai-agents-ui"
version = "0.1.0"
description = "Streamlit UI for learn-ai-agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "streamlit>=1.33,<2.0",
  "httpx>=0.27,<0.28",
  "pydantic>=2.8,<3.0",
  "python-dotenv>=1.0,<2.0",
  "motor>=3.3,<4.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["streamlit_ui"]
